requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.6
colorama==0.4.6
dj-database-url==0.5.0
Django==2.1.5
djangorestframework==3.11.2
gunicorn==20.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2-binary==2.8.6
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==5.2.0
